backend/pyproject.toml
[project]
name = "bank-statement-extraction"
version = "0.1.0"
description = "Bank statement extraction with multiple approaches"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.18.0",
    "requests>=2.31.0",
    "PyYAML>=6.0.0",
    "pdfplumber>=0.10.0",
    "Pillow>=10.0.0",
    "pdf2image>=1.16.0",
    "pytesseract>=0.3.10",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "boto3>=1.34.0",
    "alembic>=1.13.0",
    "langchain-anthropic>=0.3.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.9.0",
    "pandas>=2.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "W",
    "I",
]
ignore = [
    "E402",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["src/tests"]
pythonpath = ["."]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.basedpyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"

